# RGB Curves Luminosity mode: ~ 40% speedup