Here's a speedup for RGB Curves Luminosity mode.
Processing tome for ImProcFunctions::rgbProc went down from 1027 ms to 772 ms measured with my usual D800 file and neutral profile + RGB green curve in Luminosity mode.
That's a gross speedup of ~25%, but rgbproc does a lot of other things too which take ~394 ms, so after subtracting the 394 ms from the above mentioned values we get a net speedup of ~40% for the RGB Curves Luminosity mode part of the function.

Ingo