@Floessie 2016-08-26
:+1: for also streamlining the code. Now it's a pleasure to read.