@heckflosse 2016-08-26
I added a modification which reduces the number of atan2 calculations. Processing time is reduced to ~720 ms.