@heckflosse 2016-08-26
I added an experimental modification to equalize the effect of R, G and B curve in luminosity mode.
For reference: #3266