@michaelezra 2016-08-26
Hi Ingo, great to have such speedups. Since migration to git I've lost ability to compile rt.. If there is an exe, I 'd be glad to test:)