@heckflosse 2016-08-26
Hi Michael, my question was about the last (experimental) change (equalizing strength of rgv curves in luminosity mode). That will break backward compatibility for R and B curves in luminosity mode but will also increase usability.

Ingo