@heckflosse 2016-08-26
@michaelezra It will also change the action of red curve (all are equal after the change and almost independent on working colour space).