@michaelezra 2016-08-26
Even better;) Thank you Ingo!



-------- Original message --------From: Ingo Weyrich <notifications@github.com> Date: 8/26/16  5:27 PM  (GMT-05:00) To: Beep6581/RawTherapee <RawTherapee@noreply.github.com> Cc: Michael Ezra <michael@michaelezra.com>, Mention <mention@noreply.github.com> Subject: Re: [Beep6581/RawTherapee] RGB Curves Luminosity mode: ~ 40% speedup
  (#3404) 
@michaelezra It will also change the action of red curve (all are equal after the change and almost independent on working colour space).

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub, or mute the thread.


  
  




{"api_version":"1.0","publisher":{"api_key":"05dde50f1d1a384dd78767c55493e4bb","name":"GitHub"},"entity":{"external_key":"github/Beep6581/RawTherapee","title":"Beep6581/RawTherapee","subtitle":"GitHub repository","main_image_url":"https://cloud.githubusercontent.com/assets/143418/17495839/a5054eac-5d88-11e6-95fc-7290892c7bb5.png","avatar_image_url":"https://cloud.githubusercontent.com/assets/143418/15842166/7c72db34-2c0b-11e6-9aed-b52498112777.png","action":{"name":"Open in GitHub","url":"https://github.com/Beep6581/RawTherapee"}},"updates":{"snippets":[{"icon":"PERSON","message":"@heckflosse in #3404: @michaelezra It will also change the action of red curve (all are equal after the change and almost independent on working colour space)."}],"action":{"name":"View Pull Request","url":"https://github.com/Beep6581/RawTherapee/pull/3404#issuecomment-242856192"}}}