@heckflosse 2016-08-28
I merged to get more tests for the new behaviour of rgb red and blue curve in Luminosity mode.